# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: API documentation for Sample API
  version: 1.0.0
  contact:
    name: API Support
    url: https://www.example.com/support
    email: support@example.com
servers:
  - url: http://localhost:3005
    description: Local development server
...
paths:
...
  /aws/instances/keypair:
    post:
      summary: Create AWS EC2 keypair
      description: Create AWS EC2 Keypair used to access EC2 instances
      operationId: createKeyPair
      tags:
        - EC2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                keyName:
                  type: string
                  example: "key-pair-1"
                region:
                  type: string
                  example: "us-east-1"
              required:
                - region
                - keyName
      responses:
        "201":
          description: AWS EC2 keypair successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Key pair successfully created"
                  keyName:
                    type: string
                    example: "key-pair-1"
        "500":
          description: Internal server error